Constrainable and Buildable are not independent
ConstrainableBuildable is created to extend both the interfaces for
eliminating the need of implementing both in ObjectBuilder and
ListBuilder or other custom builders.

A trick is added for more powerful chaining. Return type of execute, set
and construct methods in ObjectBuilder is now set to ObjectBuilder. The
same is true for ListBuilder.
Therefore, new ObjectBuilder<MyClass>(...).execute(...) will now return
ObjectBuilder<MyClass> instead of Constrainable<MyClass>.
now